rename cdtp module to protocol
