+++ src/browser/transport/web_socket_connection.rs
@@ -8,7 +8,7 @@ use websocket::stream::sync::TcpStream;
 use websocket::WebSocketError;
 use websocket::{ClientBuilder, OwnedMessage};
 
-use crate::cdtp;
+use crate::protocol;
 use std::sync::Mutex;
 
 pub struct WebSocketConnection {
@@ -18,7 +18,7 @@ pub struct WebSocketConnection {
 impl WebSocketConnection {
     pub fn new(
         ws_url: &str,
-        target_messages_tx: mpsc::Sender<cdtp::Message>,
+        target_messages_tx: mpsc::Sender<protocol::Message>,
     ) -> Result<Self, Error> {
         let connection = Self::websocket_connection(&ws_url)?;
         let (websocket_receiver, sender) = connection.split()?;
@@ -43,7 +43,7 @@ impl WebSocketConnection {
 
     fn dispatch_incoming_messages(
         mut receiver: websocket::receiver::Reader<TcpStream>,
-        messages_tx: mpsc::Sender<cdtp::Message>,
+        messages_tx: mpsc::Sender<protocol::Message>,
     ) {
         for ws_message in receiver.incoming_messages() {
             match ws_message {
@@ -60,7 +60,7 @@ impl WebSocketConnection {
                 },
                 Ok(message) => {
                     if let OwnedMessage::Text(message_string) = message {
-                        if let Ok(message) = cdtp::parse_raw_message(&message_string) {
+                        if let Ok(message) = protocol::parse_raw_message(&message_string) {
                             if messages_tx.send(message).is_err() {
                                 break;
                             }
@@ -78,7 +78,10 @@ impl WebSocketConnection {
         }
 
         trace!("Sending shutdown message to message handling loop");
-        if messages_tx.send(cdtp::Message::ConnectionShutdown).is_err() {
+        if messages_tx
+            .send(protocol::Message::ConnectionShutdown)
+            .is_err()
+        {
             warn!("Couldn't send message to transport loop telling it to shut down")
         }
     }
